rough draft
Here's a quick and messy version of the check-in page. If you open index.html, fill out all the options, and click Submit, you'll see the customer-facing form. That form is approved and locked in, so we don't want to change anything about it. However, the code needs some work. I'm going to re-do it with a focus on making it easier for other stores to choose their own options down the road.